+++ ex33.py
@@ -3,9 +3,9 @@ from math import exp, cos, sin
 x1 = raw_input("X= ")
 y1 = raw_input("Y= ")
 z1 = raw_input("Z= ")
-x = int(x1)
-y = int(y1)
-z = int(z1)
+x = float(x1)
+y = float(y1)
+z = float(z1)
 
 a = (1+y)*((x+(y/((x**2)+4)))/(exp(-x-2)+(1/(x**2+4))))
 b = (1+cos(y-2))/(((x**4)/2)+(sin(z))**2)
